populate owner on comments

+++ comments/views.py
@@ -1,17 +1,17 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response 
-# from rest_framework.permissions import IsAuthenticated
+from rest_framework.permissions import IsAuthenticated
 from rest_framework import status 
-# from rest_framework.exceptions import NotFound, PermissionDenied
+from rest_framework.exceptions import NotFound, PermissionDenied
 
 from .serializers.common import CommentSerializer
 from .models import Comment 
 
 class CommentListView(APIView):
-    # permission_classes = (IsAuthenticated,)
+    permission_classes = (IsAuthenticated,)
 
     def post(self, request):
-        # request.data["owner"] = request.user.id 
+        request.data["owner"] = request.user.id 
         comment_to_create = CommentSerializer(data=request.data) 
         if comment_to_create.is_valid():
             comment_to_create.save()
@@ -25,7 +25,7 @@ class CommentDetailView(APIView):
             comment_to_delete = Comment.objects.get(pk=pk)
         except Comment.DoesNotExist:
             raise NotFound()
-        # if comment_to_delete.owner != request.user:
-        #     raise PermissionDenied()    
+        if comment_to_delete.owner != request.user:
+            raise PermissionDenied()    
         comment_to_delete.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
